refactor: RankingList 컴포넌트에 title props 추가
